move domain class from stix directory
