+++ nav-app/src/app/classes/domain.ts
@@ -1,15 +1,15 @@
-import { ServiceAuth } from '../../services/data.service';
-import { Campaign } from './campaign';
-import { DataComponent } from './data-component';
-import { Group } from './group';
-import { Matrix } from './matrix';
-import { Mitigation } from './mitigation';
-import { Note } from './note';
-import { Software } from './software';
-import { Tactic } from './tactic';
-import { Technique } from './technique';
-import { Version } from '../version';
-import { Asset } from './asset';
+import { ServiceAuth } from '../services/data.service';
+import { Campaign } from './stix/campaign';
+import { DataComponent } from './stix/data-component';
+import { Group } from './stix/group';
+import { Matrix } from './stix/matrix';
+import { Mitigation } from './stix/mitigation';
+import { Note } from './stix/note';
+import { Software } from './stix/software';
+import { Tactic } from './stix/tactic';
+import { Technique } from './stix/technique';
+import { Version } from './version';
+import { Asset } from './stix/asset';
 
 export class Domain {
     public readonly id: string; // domain ID
