+++ nav-app/src/app/classes/index.ts
@@ -1,5 +1,6 @@
 export { VersionChangelog } from './version-changelog';
 export { Version } from './version';
+export { Domain } from './domain';
 export { ContextMenuItem } from './context-menu-item';
 export { Gradient, Gcolor } from './gradient';
 export { Filter } from './filter';
