+++ nav-app/src/app/tabs/tabs.component.ts
@@ -1,7 +1,6 @@
 import { Component, ViewChild, TemplateRef, AfterViewInit, ViewEncapsulation, Input, Output, EventEmitter } from '@angular/core';
 import { DataService } from '../services/data.service';
-import { Domain } from '../classes/stix';
-import { Tab, Version, ViewModel } from '../classes';
+import { Tab, Domain, Version, ViewModel } from '../classes';
 import { ConfigService } from '../services/config.service';
 import { VersionUpgradeComponent } from '../version-upgrade/version-upgrade.component';
 import { HelpComponent } from '../help/help.component';
